+++ src/Map/Map.tsx
@@ -1,31 +1,32 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import React, { MutableRefObject, useEffect, useRef } from 'react';
+import { MutableRefObject, useEffect, useRef } from 'react';
 
-const { kakao } = window;
+declare global {
+  interface Window {
+    kakao: any;
+  }
+}
 
-const Map = () => {
+function Map() {
   const mapRef = useRef<HTMLElement | null>(null);
 
-  useEffect(() => {
-    const container = document.getElementById('myMap');
+  const initMap = () => {
+    const container = document.getElementById('map');
     const options = {
-      center: new kakao.maps.LatLng(33.450701, 126.570667),
-      level: 3,
+      center: new window.kakao.maps.LatLng(37.483034, 126.902435),
+      level: 2,
     };
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const map = new kakao.maps.Map(container as HTMLElement, options);
-    (mapRef as MutableRefObject<any>).current = map;
-  }, []);
 
-  return (
-    <div
-      id="myMap"
-      style={{
-        width: '500px',
-        height: '500px',
-      }}
-    ></div>
-  );
-};
+    const map = new window.kakao.maps.Map(container as HTMLElement, options);
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    (mapRef as MutableRefObject<any>).current = map;
+  };
+
+  useEffect(() => {
+    window.kakao.maps.load(() => initMap());
+  }, [mapRef]);
+
+  return <div id="map" style={{ width: '500px', height: '400px' }}></div>;
+}
 
 export default Map;
